Restructure persistence packages to group them instead of top level
