+++ docs/examples/kubernetes/meeseeks-configmap.yaml
