+++ docs/examples/kubernetes/meeseeks-deployment.yaml
