+++ mocks/mocks.go
@@ -12,9 +12,9 @@ import (
 	"time"
 
 	"github.com/gomeeseeks/meeseeks-box/config"
-	"github.com/gomeeseeks/meeseeks-box/db"
 	"github.com/gomeeseeks/meeseeks-box/formatter"
 	"github.com/gomeeseeks/meeseeks-box/meeseeks"
+	"github.com/gomeeseeks/meeseeks-box/persistence/db"
 
 	"github.com/sirupsen/logrus"
 )
