+++ persistence/tokens/tokens.go
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	"github.com/coreos/bbolt"
-	"github.com/gomeeseeks/meeseeks-box/db"
 	"github.com/gomeeseeks/meeseeks-box/meeseeks"
+	"github.com/gomeeseeks/meeseeks-box/persistence/db"
 	"github.com/sirupsen/logrus"
 )
