+++ persistence/tokens/tokens_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/gomeeseeks/meeseeks-box/meeseeks"
 	"github.com/gomeeseeks/meeseeks-box/mocks"
-	"github.com/gomeeseeks/meeseeks-box/tokens"
+	"github.com/gomeeseeks/meeseeks-box/persistence/tokens"
 )
 
 func Test_TokenLifecycle(t *testing.T) {
